Extract AppState state types into src/app/state/
src/app.rs grew to 6659 lines, dominating the codebase. This is the first
of three slices that splits it into a src/app/ module mirroring the
src/buffer/ convention. No behaviour change.

This slice converts src/app.rs to src/app/mod.rs and moves the 19 state
structs/enums that lived at the top of the file (InputMode, FuzzyPickerState,
SymbolPickerState, ProjectSearchState, SidebarState and friends,
CompletionState, HoverState, ReferencesListState, DiffPeekState,
ClipboardRingState, ReferenceItem, LspApprovalReason, PendingLspApproval,
LSP_SERVER_OPTIONS, LspPickerState, SidebarDrag) into six type-per-domain
files under src/app/state/. The sidebar_scroll_tests module travels with
SidebarState. A pub use in app/mod.rs preserves every crate::app::FooState
import path used by src/ui/*.rs.
