+++ src/app/mod.rs
@@ -20,7 +20,7 @@ use crate::{
         action::{Direction, EditorAction, ScrollDir},
         keybinding::KeyBindings,
     },
-    search::{SearchState, project::ProjectSearchResults},
+    search::SearchState,
     ui,
     ui::command_palette::CommandPaletteState,
     ui::editor_view::effective_gutter_width,
@@ -84,874 +84,9 @@ pub const MAX_SIDEBAR_WIDTH: u16 = 80;
 /// the cursor).
 const DOUBLE_CLICK_INTERVAL: Duration = Duration::from_millis(500);
 
-// ── Modal input mode ─────────────────────────────────────────────────────────
-
-/// Modes that capture keyboard input for a status-bar prompt.
-#[derive(Debug, Clone, PartialEq)]
-pub enum InputMode {
-    Normal,
-    /// Ctrl+G: "Go to line: {input}"
-    JumpToLine(String),
-    /// Ctrl+O: "Open: {input}"
-    OpenFilePath(String),
-    /// Ctrl+Shift+S: "Save as: {input}"
-    SaveAsPath(String),
-    /// F2 (sidebar): "Rename: {input}" — carries (original_path, current_input).
-    RenamePath(PathBuf, String),
-    /// Ctrl+Shift+N (sidebar): "New folder: {input}" — carries (parent_dir, current_input).
-    NewFolderName(PathBuf, String),
-    /// F2: "Rename: {input}" (LSP rename symbol)
-    Rename(String),
-    /// Git dialog → "Commit message: {input}". On submit, runs `git commit -m`.
-    GitCommitMessage(String),
-    /// Git dialog → "New branch: {input}". On submit, runs `git checkout -b`.
-    GitNewBranch(String),
-    /// Git dialog → "Stash message: {input}" (optional). On submit, runs `git stash push`.
-    GitStashMessage(String),
-    /// Ctrl+Alt+\ → "Shell filter (selection): {cmd}". On Enter, runs the
-    /// command via `sh -c` with the selection on stdin and replaces the
-    /// selection with the captured stdout.
-    ShellFilter(String),
-    /// "Align on character: " — `apply_align_on` is invoked with the first
-    /// non-empty character submitted (Enter takes the first char, or aborts
-    /// if empty).
-    AlignChar(String),
-    /// Ctrl+M: "Mark: " — auto-submits on the next typed alphabetic char.
-    SetMarkChar,
-    /// Ctrl+': "Jump to mark: " — auto-submits on the next typed char.
-    JumpToMarkChar,
-    /// "Record macro into slot: " — next char names the slot a–z.
-    RecordMacroChar,
-    /// "Replay macro from slot: " — next char names the slot a–z.
-    ReplayMacroChar,
-    /// Alt+': "Surround with: " — wraps the current selection in a chosen
-    /// delimiter pair on the next typed character.
-    SurroundChar,
-}
-
-impl InputMode {
-    pub fn is_normal(&self) -> bool {
-        matches!(self, InputMode::Normal)
-    }
-
-    /// Mutable access to the prompt buffer for string-carrying variants.
-    ///
-    /// Used by `handle_modal_input` to share a single InsertChar/DeleteBackward
-    /// code path across every text-entry prompt. Returns `None` for variants
-    /// that don't store an editable string (Normal, *Char variants).
-    pub fn string_mut(&mut self) -> Option<&mut String> {
-        match self {
-            InputMode::JumpToLine(s)
-            | InputMode::OpenFilePath(s)
-            | InputMode::SaveAsPath(s)
-            | InputMode::RenamePath(_, s)
-            | InputMode::NewFolderName(_, s)
-            | InputMode::Rename(s)
-            | InputMode::GitCommitMessage(s)
-            | InputMode::GitNewBranch(s)
-            | InputMode::GitStashMessage(s)
-            | InputMode::ShellFilter(s)
-            | InputMode::AlignChar(s) => Some(s),
-            InputMode::Normal
-            | InputMode::SetMarkChar
-            | InputMode::JumpToMarkChar
-            | InputMode::RecordMacroChar
-            | InputMode::ReplayMacroChar
-            | InputMode::SurroundChar => None,
-        }
-    }
-}
-
-// ── Fuzzy picker state ────────────────────────────────────────────────────────
-
-pub struct FuzzyPickerState {
-    pub query: String,
-    /// All files in the project directory (populated once on open).
-    pub all_files: Vec<PathBuf>,
-    /// Scored and sorted (score DESC) indices into `all_files`.
-    pub filtered: Vec<(u32, usize)>,
-    /// Currently highlighted row (0-based within `filtered`).
-    pub selected: usize,
-}
-
-impl FuzzyPickerState {
-    /// Build by walking the current directory with `ignore` (respects .gitignore).
-    /// `hide_git` skips the `.git` directory; `hide_dot` skips every dot-prefixed
-    /// directory (and implies `hide_git`).
-    pub fn new(hide_git: bool, hide_dot: bool) -> Self {
-        let mut all_files = Vec::new();
-        let mut builder = ignore::WalkBuilder::new(".");
-        builder.hidden(false).git_ignore(true);
-        crate::search::apply_hidden_dir_filters(&mut builder, hide_git, hide_dot);
-        for entry in builder.build().flatten() {
-            if entry.file_type().map(|ft| ft.is_file()).unwrap_or(false) {
-                // Strip the leading "./" for display clarity.
-                let p = entry.into_path();
-                let p = p.strip_prefix("./").map(PathBuf::from).unwrap_or(p);
-                all_files.push(p);
-            }
-        }
-        all_files.sort();
-        let n = all_files.len().min(200); // show first 200 unfiltered
-        let filtered = (0..n).map(|i| (0u32, i)).collect();
-        Self {
-            query: String::new(),
-            all_files,
-            filtered,
-            selected: 0,
-        }
-    }
-
-    /// Re-score the file list against the current query using nucleo.
-    pub fn update_query(&mut self, query: String) {
-        self.query = query;
-        self.selected = 0;
-
-        if self.query.is_empty() {
-            let n = self.all_files.len().min(200);
-            self.filtered = (0..n).map(|i| (0u32, i)).collect();
-            return;
-        }
-
-        use nucleo::pattern::{CaseMatching, Normalization, Pattern};
-        use nucleo::{Config, Matcher, Utf32String};
-
-        let mut matcher = Matcher::new(Config::DEFAULT);
-        let pattern = Pattern::parse(&self.query, CaseMatching::Smart, Normalization::Smart);
-
-        let mut scored: Vec<(u32, usize)> = self
-            .all_files
-            .iter()
-            .enumerate()
-            .filter_map(|(idx, path)| {
-                let s = path.to_string_lossy();
-                let haystack = Utf32String::from(s.as_ref());
-                pattern
-                    .score(haystack.slice(..), &mut matcher)
-                    .map(|sc| (sc, idx))
-            })
-            .collect();
-
-        scored.sort_by_key(|b| std::cmp::Reverse(b.0));
-        scored.truncate(200);
-        self.filtered = scored;
-    }
-
-    /// Build a picker pre-populated with an explicit path list (for recent files).
-    pub fn from_paths(paths: Vec<PathBuf>) -> Self {
-        let n = paths.len().min(200);
-        let filtered = (0..n).map(|i| (0u32, i)).collect();
-        Self {
-            query: String::new(),
-            all_files: paths,
-            filtered,
-            selected: 0,
-        }
-    }
-
-    /// Build a picker pre-populated with open buffer names (for buffer switcher).
-    /// The `all_files` list stores synthetic paths using the buffer display name.
-    pub fn from_buffers(names: Vec<(usize, String)>) -> Self {
-        let all_files: Vec<PathBuf> = names.iter().map(|(_, name)| PathBuf::from(name)).collect();
-        let n = all_files.len();
-        let filtered = (0..n).map(|i| (0u32, i)).collect();
-        Self {
-            query: String::new(),
-            all_files,
-            filtered,
-            selected: 0,
-        }
-    }
-
-    pub fn selected_path(&self) -> Option<&PathBuf> {
-        self.filtered
-            .get(self.selected)
-            .map(|(_, idx)| &self.all_files[*idx])
-    }
-
-    pub fn move_up(&mut self) {
-        if self.selected > 0 {
-            self.selected -= 1;
-        }
-    }
-
-    pub fn move_down(&mut self) {
-        if !self.filtered.is_empty() && self.selected < self.filtered.len() - 1 {
-            self.selected += 1;
-        }
-    }
-}
-
-// ── Symbol-in-file picker state ──────────────────────────────────────────
-
-/// State for the Ctrl+Shift+O symbol-in-file picker. Mirrors
-/// [`FuzzyPickerState`] but scores against symbol names rather than file
-/// paths and tracks each symbol's byte range so `Enter` can jump the cursor.
-pub struct SymbolPickerState {
-    pub query: String,
-    /// All symbols collected from the active buffer's parse tree.
-    pub all_symbols: Vec<crate::syntax::Symbol>,
-    /// Scored and sorted (score DESC) indices into `all_symbols`.
-    pub filtered: Vec<(u32, usize)>,
-    /// Currently highlighted row.
-    pub selected: usize,
-}
-
-impl SymbolPickerState {
-    /// Build with the symbols already collected from the active buffer.
-    pub fn new(symbols: Vec<crate::syntax::Symbol>) -> Self {
-        let n = symbols.len();
-        let filtered = (0..n).map(|i| (0u32, i)).collect();
-        Self {
-            query: String::new(),
-            all_symbols: symbols,
-            filtered,
-            selected: 0,
-        }
-    }
-
-    /// Re-score against the current query using nucleo. Empty query shows
-    /// every symbol in source order.
-    pub fn update_query(&mut self, query: String) {
-        self.query = query;
-        self.selected = 0;
-
-        if self.query.is_empty() {
-            let n = self.all_symbols.len();
-            self.filtered = (0..n).map(|i| (0u32, i)).collect();
-            return;
-        }
-
-        use nucleo::pattern::{CaseMatching, Normalization, Pattern};
-        use nucleo::{Config, Matcher, Utf32String};
-
-        let mut matcher = Matcher::new(Config::DEFAULT);
-        let pattern = Pattern::parse(&self.query, CaseMatching::Smart, Normalization::Smart);
-
-        let mut scored: Vec<(u32, usize)> = self
-            .all_symbols
-            .iter()
-            .enumerate()
-            .filter_map(|(idx, sym)| {
-                let haystack = Utf32String::from(sym.name.as_str());
-                pattern
-                    .score(haystack.slice(..), &mut matcher)
-                    .map(|sc| (sc, idx))
-            })
-            .collect();
-        scored.sort_by_key(|b| std::cmp::Reverse(b.0));
-        self.filtered = scored;
-    }
-
-    pub fn selected_symbol(&self) -> Option<&crate::syntax::Symbol> {
-        self.filtered
-            .get(self.selected)
-            .map(|(_, idx)| &self.all_symbols[*idx])
-    }
-
-    pub fn move_up(&mut self) {
-        if self.selected > 0 {
-            self.selected -= 1;
-        }
-    }
-
-    pub fn move_down(&mut self) {
-        if !self.filtered.is_empty() && self.selected < self.filtered.len() - 1 {
-            self.selected += 1;
-        }
-    }
-}
-
-// ── Project search overlay state ───────────────────────────────────────────
-
-/// State for the project-wide search-and-replace overlay (Ctrl+Shift+F).
-pub struct ProjectSearchState {
-    pub query: String,
-    pub replace_text: String,
-    pub is_regex: bool,
-    pub case_sensitive: bool,
-    pub show_replace: bool,
-    pub focus_replace: bool,
-    pub results: ProjectSearchResults,
-    /// Index into `results.matches` for the highlighted row.
-    pub selected: usize,
-}
-
-impl ProjectSearchState {
-    pub fn new() -> Self {
-        Self {
-            query: String::new(),
-            replace_text: String::new(),
-            is_regex: false,
-            case_sensitive: false,
-            show_replace: false,
-            focus_replace: false,
-            results: ProjectSearchResults::default(),
-            selected: 0,
-        }
-    }
-
-    pub fn move_up(&mut self) {
-        if self.selected > 0 {
-            self.selected -= 1;
-        }
-    }
-
-    pub fn move_down(&mut self) {
-        if !self.results.matches.is_empty() && self.selected + 1 < self.results.matches.len() {
-            self.selected += 1;
-        }
-    }
-}
-
-impl Default for ProjectSearchState {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-// ── Sidebar / file tree ────────────────────────────────────────────────────
-
-#[derive(Debug, Clone)]
-/// Tracks the state of a sidebar delete confirmation.
-pub enum ConfirmDelete {
-    /// Deleting a file — waiting for Y/N.
-    File(PathBuf),
-    /// Deleting a directory — first step, waiting for Y/N.
-    Dir(PathBuf),
-    /// Deleting a directory — user pressed Y, now waiting for Enter to confirm.
-    DirConfirmed(PathBuf),
-}
-
-/// Tracks a file that has been cut or copied in the sidebar.
-pub struct SidebarClipboard {
-    pub path: PathBuf,
-    pub is_cut: bool, // true = move, false = copy
-}
-
-pub struct TreeEntry {
-    pub path: PathBuf,
-    pub depth: usize,
-    pub is_dir: bool,
-    pub expanded: bool,
-}
-
-pub struct SidebarState {
-    pub entries: Vec<TreeEntry>,
-    pub selected: usize,
-    /// Index of the first visible entry. Independent from `selected`:
-    /// scroll-wheel and resize move this without touching `selected`,
-    /// while keyboard navigation calls `ensure_selected_visible` to keep
-    /// the selection on screen.
-    pub scroll_offset: usize,
-    pub root: PathBuf,
-}
-
-impl SidebarState {
-    pub fn new() -> Self {
-        let root = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
-        let mut state = Self {
-            entries: Vec::new(),
-            selected: 0,
-            scroll_offset: 0,
-            root: root.clone(),
-        };
-        state.load_root();
-        state
-    }
-
-    /// Scroll the visible window by `delta_lines` rows (positive = down).
-    /// Selection is unchanged. Clamped to `[0, max_scroll]` where
-    /// `max_scroll` keeps at least one row visible when `viewport_rows > 0`.
-    pub fn scroll_by(&mut self, delta_lines: isize, viewport_rows: usize) {
-        let max_scroll = if viewport_rows == 0 || self.entries.len() <= viewport_rows {
-            0
-        } else {
-            self.entries.len() - viewport_rows
-        };
-        let new = (self.scroll_offset as isize + delta_lines).max(0) as usize;
-        self.scroll_offset = new.min(max_scroll);
-    }
-
-    /// Adjust `scroll_offset` so the currently-selected entry is on-screen.
-    /// No-op when `viewport_rows == 0`.
-    pub fn ensure_selected_visible(&mut self, viewport_rows: usize) {
-        if viewport_rows == 0 {
-            return;
-        }
-        if self.selected < self.scroll_offset {
-            self.scroll_offset = self.selected;
-        } else if self.selected >= self.scroll_offset + viewport_rows {
-            self.scroll_offset = self.selected + 1 - viewport_rows;
-        }
-    }
-
-    /// Clamp `scroll_offset` so it doesn't point past the end of `entries`.
-    /// Called after operations that shrink or change the entry list.
-    fn clamp_scroll(&mut self) {
-        let max = self.entries.len().saturating_sub(1);
-        if self.scroll_offset > max {
-            self.scroll_offset = max;
-        }
-    }
-
-    /// Load the top-level entries of the root directory.
-    fn load_root(&mut self) {
-        self.entries.clear();
-        // Root node is always present and always expanded; it cannot be collapsed.
-        self.entries.push(TreeEntry {
-            path: self.root.clone(),
-            depth: 0,
-            is_dir: true,
-            expanded: true,
-        });
-        self.entries_from_dir(&self.root.clone(), 1, true);
-    }
-
-    /// Append entries for a directory at `depth`. If `expand` is false, only
-    /// add the directory entry itself (collapsed).
-    fn entries_from_dir(&mut self, dir: &PathBuf, depth: usize, _expand: bool) {
-        let mut children: Vec<(PathBuf, bool)> = Vec::new();
-        if let Ok(read_dir) = std::fs::read_dir(dir) {
-            for entry in read_dir.flatten() {
-                let path = entry.path();
-                let is_dir = path.is_dir();
-                children.push((path, is_dir));
-            }
-        }
-        // Sort: dirs first, then files, both alphabetically.
-        children.sort_by(|a, b| {
-            b.1.cmp(&a.1)
-                .then_with(|| a.0.file_name().cmp(&b.0.file_name()))
-        });
-        for (path, is_dir) in children {
-            self.entries.push(TreeEntry {
-                path,
-                depth,
-                is_dir,
-                expanded: false,
-            });
-        }
-    }
-
-    /// Toggle expand/collapse of the selected directory entry.
-    #[allow(dead_code)]
-    pub fn toggle_selected(&mut self) {
-        let idx = self.selected;
-        if idx >= self.entries.len() {
-            return;
-        }
-        let entry = &self.entries[idx];
-        if !entry.is_dir {
-            return;
-        }
-        // Root cannot be collapsed.
-        if entry.path == self.root {
-            return;
-        }
-        if entry.expanded {
-            // Collapse: remove all children (entries with depth > entry.depth
-            // that follow this entry and have higher depth).
-            let depth = entry.depth;
-            self.entries[idx].expanded = false;
-            let start = idx + 1;
-            let end = self.entries[start..]
-                .iter()
-                .position(|e| e.depth <= depth)
-                .map(|p| start + p)
-                .unwrap_or(self.entries.len());
-            self.entries.drain(start..end);
-        } else {
-            // Expand: load children and insert after this entry.
-            let dir = self.entries[idx].path.clone();
-            let depth = self.entries[idx].depth;
-            self.entries[idx].expanded = true;
-            let mut children: Vec<TreeEntry> = Vec::new();
-            let mut tmp = Self {
-                entries: Vec::new(),
-                selected: 0,
-                scroll_offset: 0,
-                root: dir.clone(),
-            };
-            tmp.entries_from_dir(&dir, depth + 1, false);
-            children.extend(tmp.entries);
-            let insert_at = idx + 1;
-            for (i, child) in children.into_iter().enumerate() {
-                self.entries.insert(insert_at + i, child);
-            }
-        }
-        self.clamp_scroll();
-    }
-
-    /// Collapse the directory entry at `idx` (if expanded), removing its children.
-    fn collapse_at(&mut self, idx: usize) {
-        if idx >= self.entries.len() || !self.entries[idx].is_dir || !self.entries[idx].expanded {
-            return;
-        }
-        // Root cannot be collapsed.
-        if self.entries[idx].path == self.root {
-            return;
-        }
-        let depth = self.entries[idx].depth;
-        self.entries[idx].expanded = false;
-        let start = idx + 1;
-        let end = self.entries[start..]
-            .iter()
-            .position(|e| e.depth <= depth)
-            .map(|p| start + p)
-            .unwrap_or(self.entries.len());
-        self.entries.drain(start..end);
-        self.clamp_scroll();
-    }
-
-    /// Move selection to the nearest ancestor directory and collapse it.
-    /// Does nothing if the selected entry is already at depth 0.
-    pub fn move_to_parent_and_collapse(&mut self) {
-        let idx = self.selected;
-        let depth = match self.entries.get(idx) {
-            Some(e) => e.depth,
-            None => return,
-        };
-        if depth == 0 {
-            return;
-        }
-        if let Some(parent_idx) = self.entries[..idx]
-            .iter()
-            .rposition(|e| e.depth == depth - 1)
-        {
-            self.selected = parent_idx;
-            self.collapse_at(parent_idx);
-        }
-    }
-
-    /// Expand the directory entry at `idx` without affecting `self.selected`.
-    fn expand_dir_at(&mut self, idx: usize) {
-        if self.entries[idx].expanded || !self.entries[idx].is_dir {
-            return;
-        }
-        let dir = self.entries[idx].path.clone();
-        let depth = self.entries[idx].depth;
-        self.entries[idx].expanded = true;
-        let mut tmp = Self {
-            entries: Vec::new(),
-            selected: 0,
-            scroll_offset: 0,
-            root: dir.clone(),
-        };
-        tmp.entries_from_dir(&dir, depth + 1, false);
-        let insert_at = idx + 1;
-        for (i, child) in tmp.entries.into_iter().enumerate() {
-            self.entries.insert(insert_at + i, child);
-        }
-    }
-
-    /// Expand all ancestor directories leading to `target` and select it.
-    /// Does nothing if `target` is not under `self.root`.
-    pub fn expand_to_path(&mut self, target: &std::path::Path) {
-        // Relative paths (e.g. from the fuzzy picker) are resolved against root.
-        let abs_target = if target.is_absolute() {
-            target.to_path_buf()
-        } else {
-            self.root.join(target)
-        };
-        let Ok(relative) = abs_target.strip_prefix(&self.root) else {
-            return;
-        };
-        let mut current = self.root.clone();
-        let components: Vec<_> = relative.components().collect();
-        for (i, component) in components.iter().enumerate() {
-            current = current.join(component);
-            let is_last = i == components.len() - 1;
-            if let Some(idx) = self.entries.iter().position(|e| e.path == current) {
-                if is_last {
-                    self.selected = idx;
-                } else {
-                    self.expand_dir_at(idx);
-                }
-            } else {
-                break;
-            }
-        }
-    }
-
-    #[allow(dead_code)]
-    pub fn move_up(&mut self) {
-        if self.selected > 0 {
-            self.selected -= 1;
-        }
-    }
-
-    #[allow(dead_code)]
-    pub fn move_down(&mut self) {
-        if self.selected + 1 < self.entries.len() {
-            self.selected += 1;
-        }
-    }
-
-    #[allow(dead_code)]
-    pub fn selected_path(&self) -> Option<&PathBuf> {
-        self.entries.get(self.selected).map(|e| &e.path)
-    }
-
-    /// Returns true if the root directory itself is currently selected.
-    fn root_is_selected(&self) -> bool {
-        self.entries
-            .get(self.selected)
-            .map(|e| e.path == self.root)
-            .unwrap_or(false)
-    }
-
-    /// Reload the sidebar, preserving expanded directories and selection by path.
-    pub fn refresh(&mut self) {
-        let expanded: Vec<PathBuf> = self
-            .entries
-            .iter()
-            .filter(|e| e.is_dir && e.expanded)
-            .map(|e| e.path.clone())
-            .collect();
-        let old_path = self.selected_path().cloned();
-        let old_selected = self.selected;
-        self.load_root();
-        // Re-expand previously expanded directories.
-        for path in &expanded {
-            if let Some(idx) = self
-                .entries
-                .iter()
-                .position(|e| &e.path == path && e.is_dir)
-                && !self.entries[idx].expanded
-            {
-                self.selected = idx;
-                self.toggle_selected();
-            }
-        }
-        // Restore selection by path if possible, otherwise clamp the old index.
-        if let Some(ref old) = old_path
-            && let Some(idx) = self.entries.iter().position(|e| &e.path == old)
-        {
-            self.selected = idx;
-            self.clamp_scroll();
-            return;
-        }
-        self.selected = old_selected.min(self.entries.len().saturating_sub(1));
-        self.clamp_scroll();
-    }
-}
-
-/// Generate a copy target path with a `-N` suffix (before the extension).
-/// Returns `None` if no suitable name can be found within 1000 attempts.
-fn copy_target_path(source: &std::path::Path, dest_dir: &std::path::Path) -> Option<PathBuf> {
-    let stem = source
-        .file_stem()
-        .and_then(|s| s.to_str())
-        .unwrap_or("file");
-    let ext = source.extension().and_then(|e| e.to_str());
-    for n in 1..1000 {
-        let name = match ext {
-            Some(e) => format!("{}-{}.{}", stem, n, e),
-            None => format!("{}-{}", stem, n),
-        };
-        let candidate = dest_dir.join(&name);
-        if !candidate.exists() {
-            return Some(candidate);
-        }
-    }
-    None
-}
-
-// ── LSP picker state ─────────────────────────────────────────────────────────
-
-/// State for the code completion popup.
-pub struct CompletionState {
-    /// All items received from the server.
-    pub items: Vec<CompletionItemEntry>,
-    /// Indices into `items` after prefix filtering.
-    pub filtered: Vec<usize>,
-    /// Currently highlighted row in `filtered`.
-    pub selected: usize,
-    /// Byte offset where completion was triggered (start of the prefix).
-    pub anchor_byte: usize,
-    /// Line of the trigger position (for popup positioning).
-    #[allow(dead_code)]
-    pub anchor_line: usize,
-    /// Display column of the trigger position.
-    #[allow(dead_code)]
-    pub anchor_col: usize,
-}
-
-/// A single completion item (simplified from LSP).
-pub struct CompletionItemEntry {
-    pub label: String,
-    pub detail: Option<String>,
-    pub insert_text: String,
-    pub filter_text: String,
-    pub kind_label: &'static str,
-}
-
-impl CompletionState {
-    pub fn new(anchor_byte: usize, anchor_line: usize, anchor_col: usize) -> Self {
-        Self {
-            items: Vec::new(),
-            filtered: Vec::new(),
-            selected: 0,
-            anchor_byte,
-            anchor_line,
-            anchor_col,
-        }
-    }
-
-    /// Re-filter items against the typed prefix.
-    pub fn filter(&mut self, prefix: &str) {
-        let lower_prefix = prefix.to_lowercase();
-        self.filtered = self
-            .items
-            .iter()
-            .enumerate()
-            .filter(|(_, item)| item.filter_text.to_lowercase().contains(&lower_prefix))
-            .map(|(i, _)| i)
-            .collect();
-        if self.selected >= self.filtered.len() {
-            self.selected = 0;
-        }
-    }
-
-    /// Get the currently selected item, if any.
-    pub fn selected_item(&self) -> Option<&CompletionItemEntry> {
-        self.filtered
-            .get(self.selected)
-            .and_then(|&i| self.items.get(i))
-    }
-}
-
-/// State for the hover info popup.
-pub struct HoverState {
-    pub content: String,
-    #[allow(dead_code)]
-    pub anchor_line: usize,
-    #[allow(dead_code)]
-    pub anchor_col: usize,
-}
-
-/// State for the references list overlay.
-pub struct ReferencesListState {
-    pub items: Vec<ReferenceItem>,
-    pub selected: usize,
-}
-
-/// State for the inline diff-peek float (`Alt+H`). Lists the HEAD lines for
-/// the hunk under the cursor.
-pub struct DiffPeekState {
-    pub head_lines: Vec<String>,
-    /// Cursor line at the time the peek was opened — used to anchor the float
-    /// near the cursor.
-    #[allow(dead_code)]
-    pub anchor_line: usize,
-}
-
-/// State for the clipboard-ring picker overlay (`Ctrl+Shift+V`).
-pub struct ClipboardRingState {
-    /// Snapshot of the ring at the time the overlay was opened. Most recent
-    /// entry first. Mutated only via `selected`; the underlying
-    /// `ClipboardManager` is not touched until the user confirms a pick.
-    pub entries: Vec<String>,
-    pub selected: usize,
-}
-
-/// A single reference location.
-pub struct ReferenceItem {
-    pub path: PathBuf,
-    pub line: usize,
-    pub col: usize,
-    pub context: String,
-}
-
-/// Why the user is being prompted to approve an LSP binary.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum LspApprovalReason {
-    /// We've never seen this binary before.
-    FirstLaunch,
-    /// We have an entry for this path, but the hash on disk has changed.
-    BinaryChanged { previous_hash: String },
-}
-
-/// State for the LSP-binary approval overlay.
-#[derive(Debug, Clone)]
-pub struct PendingLspApproval {
-    /// Server identifier from the active LSP config (e.g. `"rust-analyzer"`).
-    pub server_name: String,
-    /// Raw command from the config (may be a bare name or a path).
-    pub command: String,
-    /// Command-line args from the config.
-    pub args: Vec<String>,
-    /// Path returned by `which` (or the absolute path the user gave).
-    pub display_path: PathBuf,
-    /// Canonicalized path that gets hashed and recorded in the trust store.
-    pub canonical_path: PathBuf,
-    /// SHA-256 of the binary contents, lowercase hex.
-    pub hash: String,
-    /// What triggered the prompt.
-    pub reason: LspApprovalReason,
-}
-
-/// Built-in LSP server definitions the user can choose from.
-/// Sorted alphabetically by language display name.
-pub const LSP_SERVER_OPTIONS: &[(&str, &str, &str, &[&str])] = &[
-    // (language display, server key written to lsp.toml, command, args)
-    ("C#", "omnisharp", "omnisharp", &["-lsp"]),
-    ("C/C++", "clangd", "clangd", &[]),
-    ("Go", "gopls", "gopls", &["serve"]),
-    ("Java", "jdtls", "jdtls", &[]),
-    ("Kotlin", "kotlin-lsp", "kotlin-lsp", &[]),
-    ("Lua", "lua-language-server", "lua-language-server", &[]),
-    ("Python", "pyright", "pyright-langserver", &["--stdio"]),
-    ("Rust", "rust-analyzer", "rust-analyzer", &[]),
-    (
-        "TypeScript",
-        "typescript-language-server",
-        "typescript-language-server",
-        &["--stdio"],
-    ),
-    ("Zig", "zls", "zls", &[]),
-];
-
-/// State for the LSP configuration picker overlay.
-pub struct LspPickerState {
-    /// Currently highlighted row. 0 = Disabled, 1..=N = server options.
-    pub selected: usize,
-}
-
-impl LspPickerState {
-    pub fn new(lsp_config: &crate::lsp::config::WorkspaceLspConfig) -> Self {
-        // Pre-select the currently active server, or 0 (Disabled).
-        let selected = if !lsp_config.is_active() {
-            0
-        } else {
-            lsp_config
-                .server
-                .as_deref()
-                .and_then(|key| {
-                    LSP_SERVER_OPTIONS
-                        .iter()
-                        .position(|(_, name, _, _)| *name == key)
-                        .map(|i| i + 1)
-                })
-                .unwrap_or(0)
-        };
-        Self { selected }
-    }
-
-    pub fn num_rows(&self) -> usize {
-        1 + LSP_SERVER_OPTIONS.len() // "Disabled" + servers
-    }
-}
-
-/// State of an in-progress sidebar separator-drag.
-#[derive(Debug, Clone, Copy)]
-pub struct SidebarDrag {
-    /// Column the user pressed the mouse button on (always the separator x).
-    pub start_col: u16,
-    /// Width the sidebar had when the drag began.
-    pub start_width: u16,
-}
+pub mod state;
+use state::sidebar::copy_target_path;
+pub use state::*;
 
 // ── AppState ──────────────────────────────────────────────────────────────────
 
@@ -6573,87 +5708,4 @@ fn extract_word_at(text: &str, byte_offset: usize) -> String {
         end += 1;
     }
     text[start..end].to_string()
-}
-
-#[cfg(test)]
-mod sidebar_scroll_tests {
-    use super::*;
-
-    fn dummy_entries(n: usize) -> Vec<TreeEntry> {
-        (0..n)
-            .map(|i| TreeEntry {
-                path: PathBuf::from(format!("entry{}", i)),
-                depth: 0,
-                is_dir: false,
-                expanded: false,
-            })
-            .collect()
-    }
-
-    fn state_with(n: usize) -> SidebarState {
-        SidebarState {
-            entries: dummy_entries(n),
-            selected: 0,
-            scroll_offset: 0,
-            root: PathBuf::from("/"),
-        }
-    }
-
-    #[test]
-    fn scroll_by_clamps_at_zero() {
-        let mut sb = state_with(20);
-        sb.scroll_offset = 5;
-        sb.scroll_by(-100, 10);
-        assert_eq!(sb.scroll_offset, 0);
-    }
-
-    #[test]
-    fn scroll_by_clamps_at_max_keeps_viewport_full() {
-        let mut sb = state_with(20);
-        // viewport_rows=10 → max_scroll = 20 - 10 = 10
-        sb.scroll_by(100, 10);
-        assert_eq!(sb.scroll_offset, 10);
-    }
-
-    #[test]
-    fn scroll_by_no_scroll_when_all_fit() {
-        let mut sb = state_with(5);
-        sb.scroll_by(100, 10);
-        assert_eq!(sb.scroll_offset, 0);
-    }
-
-    #[test]
-    fn scroll_by_zero_viewport_pins_offset() {
-        let mut sb = state_with(20);
-        sb.scroll_by(100, 0);
-        assert_eq!(sb.scroll_offset, 0);
-    }
-
-    #[test]
-    fn ensure_selected_visible_scrolls_down_when_below() {
-        let mut sb = state_with(50);
-        sb.selected = 30;
-        sb.scroll_offset = 0;
-        sb.ensure_selected_visible(10);
-        // selected=30, viewport=10 → scroll_offset = 30 + 1 - 10 = 21
-        assert_eq!(sb.scroll_offset, 21);
-    }
-
-    #[test]
-    fn ensure_selected_visible_scrolls_up_when_above() {
-        let mut sb = state_with(50);
-        sb.selected = 5;
-        sb.scroll_offset = 20;
-        sb.ensure_selected_visible(10);
-        assert_eq!(sb.scroll_offset, 5);
-    }
-
-    #[test]
-    fn ensure_selected_visible_no_op_when_already_in_view() {
-        let mut sb = state_with(50);
-        sb.selected = 25;
-        sb.scroll_offset = 20;
-        sb.ensure_selected_visible(10);
-        assert_eq!(sb.scroll_offset, 20);
-    }
 }
